ffi: delegate all allocations to the caller
As in this design both ulid_new_string() and ulid_write_new() would be
identical, the latter is dropped.

Due to the increased number of tests, a problem with rand() not being
reentrant became a lot more apparent.  The tests can be made to work
reliably with `--test-threads 1`, but a proper solution is still needed,
as that is a bug that affects normal use of the public APIs.

Most of the tests can also be executed in Miri with:

MIRIFLAGS="-Zmiri-disable-isolation" cargo +nightly miri test

(-Zmiri-disable-isolation is required for SystemTime::new()).