ffi: provide safer APIs with buffer size and overflow checking
Take ideas from the safe(r) C string manipulation APIs:

- like `strscpy()` (Linux Kernel[1]), `strlcpy()` (BSD[2]) and
  `strcpy_s()` (Windows[3]), take the buffer `size` as a parameter;
- like `strscpy()` and `strcpy_s()`, return a negative error code if the
  buffer is too small, since a truncated ULID could lose the uniqueness
  property very quickly;
- like `strscpy()` and `strlcpy()`, return the number of characters
  written on success (not including the terminating null byte), for
  easier management of the remaining buffer space.

Also reintroduce the terminating null byte in all cases, since not
having it is unexpected when manipulating strings in C and, thus, more
prone to mistakes.[4]  This also improves ergonomics, since most other
string APIs in C require it.

The APIs are also renamed to more accurately and uniformly represent the
semantics of writing [to strings]:

- `ulid_new_string()` -> `ulid_write_new()` [4]
- `ulid_encode()` -> `ulid_write()`

Finally, one of the C examples in the README is made intentionally more
complicated to show how the return values can (and in some cases need to)
be used.

[1] https://www.kernel.org/doc/html/latest/core-api/kernel-api.html#c.strscpy
[2] https://man.openbsd.org/strlcpy.3
[3] https://docs.microsoft.com/en-us/cpp/c-runtime-library/reference/strcpy-s-wcscpy-s-mbscpy-s?view=msvc-160
[4] This reverts a small part of fb93e0d1c375.